refactor(rupers): lazy module
