+++ src/app/app-routing.module.ts
@@ -7,12 +7,6 @@ import { WebHookComponent } from './webhook/components/webhook.component';
 import { ConceptosTurneablesComponent } from './conceptos-turneables/components/conceptos-turneables.component';
 import { MonitoreoActivacionesComponent } from './monitor-activaciones/monitoreo-activaciones.component';
 import { WebhookLogComponent } from './webhook-log/webhook-log.component';
-import { RUPElementosRupListadoComponent } from './rupers/elementos-rup-listado/elementos-rup-listado.component';
-import { ElementosRupResolver } from './rupers/elementos-rup-resolver';
-import { RUPSeccionCreateUpdateComponent } from './rupers/seccion-create-update/seccion-create-update.component';
-import { RUPPrestacionCreateUpdateComponent } from './rupers/prestacion-create-update/prestacion-create-update.component';
-import { RUPAtomoCreateUpdateComponent } from './rupers/atomo-create-update/atomo-create-update.component';
-import { RUPMoleculaCreateUpdateComponent } from './rupers/molecula-create-update/molecula-create-update.component';
 
 const appRoutes: Routes = [
     { path: '', redirectTo: '/login', pathMatch: 'full' },
@@ -27,86 +21,7 @@ const appRoutes: Routes = [
     { path: 'webhooklog', component: WebhookLogComponent, canActivate: [RoutingGuard] },
     { path: 'cda-regenerar', loadChildren: './cda/cda.module#CdaModule' },
     { path: 'novedades', loadChildren: './registro-novedades/novedades.module#NovedadesModule' },
-    {
-        path: 'elementos-rup',
-        component: RUPElementosRupListadoComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        }
-    },
-    {
-        path: 'elementos-rup/seccion/nuevo',
-        component: RUPSeccionCreateUpdateComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        },
-        pathMatch: 'full'
-    },
-    {
-        path: 'elementos-rup/seccion/:id',
-        component: RUPSeccionCreateUpdateComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        }
-    },
-    {
-        path: 'elementos-rup/prestacion/nuevo',
-        component: RUPPrestacionCreateUpdateComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        },
-        pathMatch: 'full'
-    },
-    {
-        path: 'elementos-rup/prestacion/:id',
-        component: RUPPrestacionCreateUpdateComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        }
-    },
-    {
-        path: 'elementos-rup/atomo/nuevo',
-        component: RUPAtomoCreateUpdateComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        },
-        pathMatch: 'full'
-    },
-    {
-        path: 'elementos-rup/atomo/:id',
-        component: RUPAtomoCreateUpdateComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        }
-    },
-    {
-        path: 'elementos-rup/molecula/nuevo',
-        component: RUPMoleculaCreateUpdateComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        },
-        pathMatch: 'full'
-    },
-    {
-        path: 'elementos-rup/molecula/:id',
-        component: RUPMoleculaCreateUpdateComponent,
-        canActivate: [RoutingGuard],
-        resolve: {
-            elementos: ElementosRupResolver
-        }
-    },
-
-
-
-
+    { path: 'rupers', loadChildren: './rupers/rupers.module#RupersModule' },
     { path: '**', redirectTo: '/home', pathMatch: 'full' }
 ];
