+++ src/app/rupers/components/elementos-rup-listado/elementos-rup-listado.component.html
