+++ src/app/rupers/components/elementos-rup-listado/elementos-rup-listado.component.ts
@@ -13,10 +13,10 @@ export class RUPElementosRupListadoComponent implements OnInit {
     ) { }
 
     public items = [
-        { label: 'ATOMO', handler: () => { this.goto('/elementos-rup/atomo/nuevo'); } },
-        { label: 'MOLECULA', handler: () => { this.goto('/elementos-rup/molecula/nuevo'); } },
-        { label: 'SECCION', handler: () => { this.goto('/elementos-rup/seccion/nuevo'); } },
-        { label: 'PRESTACION', handler: () => { this.goto('/elementos-rup/prestacion/nuevo'); } }
+        { label: 'ATOMO', handler: () => { this.goto('/rupers/elementos-rup/atomo/nuevo'); } },
+        { label: 'MOLECULA', handler: () => { this.goto('/rupers/elementos-rup/molecula/nuevo'); } },
+        { label: 'SECCION', handler: () => { this.goto('/rupers/elementos-rup/seccion/nuevo'); } },
+        { label: 'PRESTACION', handler: () => { this.goto('/rupers/elementos-rup/prestacion/nuevo'); } }
 
     ];
