+++ src/app/shared/snomed.service.ts
@@ -2,17 +2,9 @@ import { Observable } from 'rxjs/Observable';
 import { Injectable } from '@angular/core';
 import { Server } from '@andes/shared';
 
-/**
- * Repetido por ahora hasta modularizar mejor los componentes.
- * Deberíams hacer un ngModule de servicios o ngModule de Snomed por lo menos.
- */
-
-@Injectable()
+@Injectable({ providedIn: 'root' })
 export class SnomedService {
-    private snomedURL = '/core/term/snomed';  // URL to web api
-    private snomedURLProblema = '/core/term/snomed/problema';  // URL to web api
-    private snomedURLProcedimiento = '/core/term/snomed/procedimiento';  // URL to web api
-    private snomedURLEquipamiento = '/core/term/snomed/equipamiento';  // URL to web api
+    private snomedURL = '/core/term/snomed';
     private snomedURLexpression = '/core/term/snomed/expression';
 
     constructor(private server: Server) {
